+++ entities/party.go
@@ -12,10 +12,16 @@ type Party struct {
 }
 
 func (p Party) HexID() string {
+	if len(p.ID) == 0 {
+		return "network"
+	}
 	return hex.EncodeToString(p.ID)
 }
 
 func MakePartyID(stringID string) ([]byte, error) {
+	if stringID == "network" {
+		return []byte{}, nil
+	}
 	id, err := hex.DecodeString(stringID)
 	if err != nil {
 		return nil, fmt.Errorf("party id is not valid hex string: %v", stringID)